pyproject.toml
[tool.poetry]
name = "sdss-lvmscp"
version = "0.2.0-alpha.0"
description = "LVM spectrograph control package."
authors = ["Changgon Kim <changgonkim@khu.ac.kr>"]
maintainers = ["José Sánchez-Gallego <gallegoj@uw.edu>"]
license = "BSD-3-Clause"
readme = "README.md"
homepage = "https://github.com/sdss/lvmscp"
repository = "https://github.com/sdss/lvmscp"
documentation = "https://sdss-lvmscp.readthedocs.org"
keywords = ["astronomy", "software"]
classifiers = [
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: OS Independent",
	"Programming Language :: Python :: 3.9",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
packages = [
    { include = "lvmscp", from = "python" }
]
include = ["python/lvmscp/etc/*"]

[tool.poetry.scripts]
lvmscp = "lvmscp.__main__:lvmscp"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
sdsstools = ">=0.4.0"
click = "^8.1.2"
click-default-group = "^1.2.2"
sdss-clu = "^1.6.1"
sdss-archon = {path = "../archon", develop = true}
httpx = ">=0.18.1"
Authlib = ">=1.0.0rc1"
# sdss-cluplus = "^0.0.14"

[tool.poetry.dev-dependencies]
ipython = ">=7.11.0"
matplotlib = ">=3.1.1"
flake8 = ">=3.7.9"
doc8 = ">=0.8.0"
pytest = ">=5.2.2"
pytest-asyncio = ">=0.10.0"
pytest-cov = ">=2.8.1"
pytest-mock = ">=1.13.0"
pytest-sugar = ">=0.9.2"
isort = ">=4.3.21"
codecov = ">=2.0.15"
coverage = {version = ">=5.0", extras = ["toml"]}
ipdb = ">=0.12.3"
rstcheck = "^3.3.1"
Sphinx = ">=3.0.0"
sphinx_bootstrap_theme = ">=0.4.12"
black = ">=22.3.0"
sphinx-click = "^3.0.1"

[tool.black]
line-length = 88
target-version = ['py39']
fast = true

[tool.isort]
line_length = 79
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "SDSS", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"
known_first_party = "lvmscp"
known_sdss = ["sdsstools"]
balanced_wrapping = true
include_trailing_comma = false
lines_after_imports = 2
use_parentheses = true

[tool.pytest.ini_options]
addopts = "--cov lvmscp --cov-report xml --cov-report html --cov-report term"

[tool.coverage.run]
branch = true
include = ["python/lvmscp/*"]
omit = [
    "python/lvmscp/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
]

[build-system]
requires = ["poetry-core>=1.1.0a7"]
build-backend = "poetry.core.masonry.api"
